tests/requirements/base.txt
django-filer>=1.5.0
html5lib>=1
tox
coverage
isort
flake8
flake8-pyproject
pyflakes>=2.1
django-test-migrations
pytest
pytest-django
-e .
